Updated pageserver.py
I implemented file checking logic into the function respond() that takes in the users request and checks if the file that was requested either exists in the CWD, does exist in the CWD or if the file name contains illegal characters such as "~" or "..". in main() the CWD is changed according to the specified direcotyr to be searched specified by DOCROOT found in credentials.ini (which was submitted via canvas and not pushed). If the user then requests a specific file page and  If the file does not exist, then STATUS_NOT_FOUND is transmitted to the page along with error code 404 being displayed. If the file name contains illegal characters, then the STATUS_FORBIDDEN is transmitted and error code 403 is transmitted to the page seen by the user.